+++ src/views/index.js
@@ -1,6 +1,8 @@
 export * from './account'
 export * from './branding'
-export * from './apps'
+export * from './workspaces'
 export * from './about'
 export * from './not-found'
 export * from './search'
+export * from './active'
+export * from './available'
